Extract blockchain_options into shared cli code
